refactor: Create a support module for the tests
